Create: Criado Services

+++ src/main/java/matheus/api/api/Repository/ManutencaoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ManutencaoRepository extends JpaRepository<Manutencao, Long>{
+public interface ManutencaoRepository extends JpaRepository<Manutencao, Integer>{
     
 }
